+++ NWN.Anvil/src/main/API/Events/Native/CreatureEvents/CheckProficiencyOverride.cs
@@ -1,0 +1,9 @@
+namespace Anvil.API.Events
+{
+  public enum CheckProficiencyOverride
+  {
+    None = 0,
+    HasProficiency,
+    NoProficiency,
+  }
+}